move erigon namespace rpc methods to eth
