+++ cmd/rpcdaemon/commands/erigon_receipts.go
@@ -8,8 +8,9 @@ import (
 	"github.com/ledgerwatch/erigon/core/types"
 )
 
+// disabled for now, subset of eth_getLogs
 // GetLogsByHash implements erigon_getLogsByHash. Returns an array of arrays of logs generated by the transactions in the block given by the block's hash.
-func (api *ErigonImpl) GetLogsByHash(ctx context.Context, hash common.Hash) ([][]*types.Log, error) {
+func (api *APIImpl) GetLogsByHash(ctx context.Context, hash common.Hash) ([][]*types.Log, error) {
 	tx, err := api.db.BeginRo(ctx)
 	if err != nil {
 		return nil, err
