+++ cmd/rpcdaemon/commands/eth_block.go
@@ -300,3 +300,52 @@ func (api *APIImpl) GetBlockTransactionCountByHash(ctx context.Context, blockHas
 	n := hexutil.Uint(txAmount)
 	return &n, nil
 }
+
+// GetHeaderByNumber implements eth_getHeaderByNumber. Returns a block's header given a block number ignoring the block's transaction and uncle list (may be faster).
+func (api *APIImpl) GetHeaderByNumber(ctx context.Context, blockNumber rpc.BlockNumber) (*types.Header, error) {
+	// Pending block is only known by the miner
+	if blockNumber == rpc.PendingBlockNumber {
+		block := api.pendingBlock()
+		if block == nil {
+			return nil, nil
+		}
+		return block.Header(), nil
+	}
+
+	tx, err := api.db.BeginRo(ctx)
+	if err != nil {
+		return nil, err
+	}
+	defer tx.Rollback()
+
+	blockNum, err := getBlockNumber(blockNumber, tx)
+	if err != nil {
+		return nil, err
+	}
+
+	header := rawdb.ReadHeaderByNumber(tx, blockNum)
+	if header == nil {
+		return nil, fmt.Errorf("block header not found: %d", blockNum)
+	}
+
+	return header, nil
+}
+
+// GetHeaderByHash implements eth_getHeaderByHash. Returns a block's header given a block's hash.
+func (api *APIImpl) GetHeaderByHash(ctx context.Context, hash common.Hash) (*types.Header, error) {
+	tx, err := api.db.BeginRo(ctx)
+	if err != nil {
+		return nil, err
+	}
+	defer tx.Rollback()
+
+	header, err := rawdb.ReadHeaderByHash(tx, hash)
+	if err != nil {
+		return nil, err
+	}
+	if header == nil {
+		return nil, fmt.Errorf("block header not found: %s", hash.String())
+	}
+
+	return header, nil
+}
